+++ src/commands/history.rs
@@ -1,17 +1,17 @@
-//! Transcription history viewer.
+//! Transcription history view.
 //!
 //! Displays and manages transcription history with copy-to-clipboard functionality.
 
 use crate::clipboard::copy_to_clipboard;
-use crate::history::{HistoryManager, HistoryViewer};
+use crate::history::{HistoryManager, HistoryView};
 
-/// Displays the transcription history viewer with copy-to-clipboard functionality.
+/// Displays the transcription history view with copy-to-clipboard functionality.
 ///
 /// # Errors
 /// - If data directory cannot be determined
 /// - If history manager fails to load transcriptions
 pub async fn handle_history() -> Result<(), anyhow::Error> {
-    tracing::info!("=== ostt History Viewer ===");
+    tracing::info!("=== ostt History View ===");
 
     let data_dir = dirs::home_dir()
         .ok_or_else(|| anyhow::anyhow!("Could not determine home directory"))?
@@ -27,18 +27,18 @@ pub async fn handle_history() -> Result<(), anyhow::Error> {
         return Ok(());
     }
 
-    let mut viewer = HistoryViewer::new(entries)?;
+    let mut view = HistoryView::new(entries)?;
 
-    match viewer.run()? {
+    match view.run()? {
         Some(selected_text) => {
             copy_to_clipboard(&selected_text)?;
             tracing::info!("Selected transcription copied to clipboard");
         }
         None => {
-            tracing::debug!("History viewer exited without selection");
+            tracing::debug!("History view exited without selection");
         }
     }
 
-    tracing::debug!("History viewer closed");
+    tracing::debug!("History view closed");
     Ok(())
 }
