+++ src/commands/keywords.rs
@@ -2,7 +2,7 @@
 //!
 //! Orchestrates the keywords management UI and storage.
 
-use crate::keywords::{KeywordsManager, KeywordsViewer};
+use crate::keywords::{KeywordsManager, KeywordsView};
 use anyhow::Result;
 use dirs;
 
@@ -17,8 +17,8 @@ pub async fn handle_keywords() -> Result<()> {
 
     let mut manager = KeywordsManager::new(&config_dir)?;
 
-    let mut viewer = KeywordsViewer::new(manager.load_keywords()?)?;
-    viewer.run(&mut manager)?;
+    let mut view = KeywordsView::new(manager.load_keywords()?)?;
+    view.run(&mut manager)?;
 
     Ok(())
 }
