+++ src/commands/model.rs
@@ -1,5 +1,5 @@
-use crate::model::ModelSelector;
+use crate::model::ModelView;
 
 pub async fn handle_model() -> anyhow::Result<()> {
-    ModelSelector::new()?.run().await
+    ModelView::new()?.run().await
 }
