+++ src/commands/transcribe.rs
@@ -116,8 +116,8 @@ pub async fn handle_transcribe(
                 ));
             }
 
-            match process::picker::show_action_picker(&config_data.process.actions)? {
-                process::picker::PickerResult::Selected(selected_id) => {
+            match process::process_view::show_action_picker(&config_data.process.actions)? {
+                process::process_view::PickerResult::Selected(selected_id) => {
                     let action = config_data
                         .process
                         .get_action(&selected_id)
@@ -139,7 +139,7 @@ pub async fn handle_transcribe(
                         }
                     }
                 }
-                process::picker::PickerResult::Cancelled => {
+                process::process_view::PickerResult::Cancelled => {
                     // Cancelled — fall through to output raw transcription
                     trimmed_text
                 }
