+++ src/process/mod.rs
@@ -2,7 +2,7 @@ pub mod ai;
 pub mod bash;
 pub mod execute;
 pub mod input;
-pub mod picker;
+pub mod process_view;
 
 pub use execute::execute_action;
 pub use execute::execute_action_with_animation;
