+++ src/recording/mod.rs
@@ -6,10 +6,10 @@
 pub mod audio;
 pub mod ffmpeg;
 pub mod recording_history;
-pub mod ui;
+pub mod ostt_tui;
 pub mod visualizations;
 
 pub use audio::AudioRecorder;
 pub use ffmpeg::find_ffmpeg;
 pub use recording_history::RecordingHistory;
-pub use ui::{OsttTui, PickerEvent, RecordingCommand};
+pub use ostt_tui::{OsttTui, PickerEvent, RecordingCommand};
