+++ src/ui/components/app_layout.rs
@@ -1,5 +1,4 @@
 use ratatui::layout::{Alignment, Constraint, Layout, Rect};
-use ratatui::style::{Color, Style};
 use ratatui::widgets::{Block, Padding, Paragraph};
 use ratatui::Frame;
 
@@ -37,25 +36,4 @@ pub fn render_app_layout(frame: &mut Frame<'_>, area: Rect) -> AppLayout {
         footer,
         full: area,
     }
-}
-
-pub fn render_footer(frame: &mut Frame<'_>, area: Rect, text: &'static str) {
-    frame.render_widget(
-        Paragraph::new(text)
-            .alignment(Alignment::Center)
-            .style(Style::default().fg(Color::White).bg(Color::DarkGray)),
-        area,
-    );
-}
-
-pub fn render_title(frame: &mut Frame<'_>, area: Rect, title: &str) {
-    let label = format!(" {title} ");
-    frame.render_widget(
-        Paragraph::new(label.clone()).style(Style::default().fg(Color::White).bg(Color::Blue)),
-        Rect {
-            width: label.len() as u16,
-            height: 1,
-            ..area
-        },
-    );
 }
